requirement.txt
python-whois
requests
python-whois
shodan
dill
pexpect
pysocks
paramiko
bs4
prettytable
httpx
